refactor: Breakout layout components
