+++ web-app/src/components/UserAvatar.tsx
@@ -1,0 +1,12 @@
+import { Avatar } from "@mantine/core";
+
+export const UserAvatar = ({ avatar }: { avatar?: string | null }) =>
+  avatar ? (
+    <Avatar
+      mx="auto"
+      src={avatar}
+      alt="User Avatar"
+      radius="xl"
+      sx={{ placeSelf: "center" }}
+    />
+  ) : null;
